refactor: move monitor and plugin related method from BrokerController to BrokerServiceManager
